requirements.txt
tabulate
pandas
sqlite3
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
